requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
babel==2.17.0
billiard==4.2.4
boto3==1.42.15
botocore==1.42.15
celery==5.6.0
certifi
cffi
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.9
django-allauth==65.13.1
django-anymail==14.0
django-appconf==1.2.0
django-axes==8.1.0
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-classy-tags==4.1.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-grappelli==4.0.3
django-guardian==3.2.0
django-haystack==3.3.0
django-health-check==3.20.8
django-imagekit==6.0.0
django-js-asset==3.1.2
django-localflavor==5.0
django-model-utils==5.0.0
django-money==3.5.4
django-mptt==0.18.0
django-multiselectfield==1.0.1
django-oauth-toolkit==3.1.0
django-parler==2.3
django-polymorphic==4.5.0
django-redis==6.0.0
django-sekizai==4.1.0
django-silk==5.4.3
django-storages==1.14.6
django-taggit==6.1.0
django-timezone-field==7.2.1
django-tinymce==5.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
jmespath==1.0.1
jsonschema==4.25.1
jwcrypto==1.5.6
kombu==5.6.1
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pilkit==3.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-moneyed==3.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.4
pytest==9.0.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-stdnum==2.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
sentry-sdk==2.48.0
six==1.17.0
sqlparse==0.5.5
stripe==14.1.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14